requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
flake8==7.3.0
flake8-isort==7.0.0
idna==3.11
isort==7.0.0
mccabe==0.7.0
pep8-naming==0.15.1
pycodestyle==2.14.0
pyflakes==3.4.0
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.5.0
